Fixing math.Sqrt float32/float64 conversion in speed calculation

+++ internal/core/ecs/components/physics.go
@@ -81,7 +81,7 @@ func (p *PhysicsComponent) ApplySpeedLimit() {
 		return
 	}
 
-	speed := math.Sqrt(p.Velocity.X*p.Velocity.X + p.Velocity.Y*p.Velocity.Y)
+	speed := math.Sqrt(float64(p.Velocity.X*p.Velocity.X + p.Velocity.Y*p.Velocity.Y))
 	if speed > p.MaxSpeed {
 		// Normalize and scale to max speed
 		scale := p.MaxSpeed / speed
